fix: avoid gRPC FTS file modification (#2736)
* fix: avoid gRPC FTS file modification

* is_file_on_remote

* update message

* use warnings

* use warnings 1

+++ src/ansys/fluent/core/utils/file_transfer_service.py
@@ -6,6 +6,7 @@ import pathlib
 import random
 import shutil
 from typing import Any, Callable, List, Optional, Protocol, Union  # noqa: F401
+import warnings
 
 from alive_progress import alive_bar
 
@@ -126,10 +127,15 @@ def _get_files(
         files = []
         for file in file_name:
             file_path_check = os.path.join(path, os.path.basename(file))
+            logger.debug(f"\n file_path_check = {file_path_check} \n")
+            logger.debug(
+                f"\n is_file_path_check = {os.path.isfile(file_path_check)} \n"
+            )
             if os.path.isfile(file_path_check):
                 files.append(file_path_check)
             else:
                 files.append(file)
+        logger.debug(f"\n files = {files} \n")
     return files
 
 
@@ -229,7 +235,13 @@ class RemoteFileTransferStrategy(FileTransferStrategy):
         files = _get_files(file_name, self.host_mount_path)
         if self.client:
             for file in files:
-                if os.path.isfile(file):
+                is_file_on_remote = self.file_exists_on_remote(os.path.basename(file))
+                if is_file_on_remote:
+                    warnings.warn(
+                        f"\n{file} with the same name exists at the remote location.\n",
+                        UserWarning,
+                    )
+                elif os.path.isfile(file) and not is_file_on_remote:
                     self.client.upload_file(
                         local_filename=file,
                         remote_filename=(
@@ -257,7 +269,9 @@ class RemoteFileTransferStrategy(FileTransferStrategy):
         if self.client:
             for file in files:
                 if os.path.isfile(file):
-                    print(f"\nFile already exists. File path:\n{file}\n")
+                    warnings.warn(
+                        f"\nFile already exists. File path:\n{file}\n", UserWarning
+                    )
                 else:
                     self.client.download_file(
                         remote_filename=os.path.basename(file),
@@ -393,7 +407,10 @@ class PimFileTransferService:
                             )
                             bar()
                         else:
-                            print(f"\n{file} already uploaded.\n")
+                            warnings.warn(
+                                f"\n{file} with the same name exists at the remote location.\n",
+                                UserWarning,
+                            )
                     elif not self.file_service.file_exist(os.path.basename(file)):
                         raise FileNotFoundError(f"{file} does not exist.")
 
@@ -439,7 +456,9 @@ class PimFileTransferService:
             with alive_bar(len(files), title="Downloading...") as bar:
                 for file in files:
                     if os.path.isfile(file):
-                        print(f"\nFile already exists. File path:\n{file}\n")
+                        warnings.warn(
+                            f"\nFile already exists. File path:\n{file}\n", UserWarning
+                        )
                     else:
                         self.download_file(
                             file_name=os.path.basename(file),
